requirements.txt
# No external dependencies required
# tkinter is included with Python by default 
Pillow 